renamed reducers.ts to actions.ts
